Move social auth demo database config to app db
