+++ demos/social-auth/app/db.test.ts
@@ -1,14 +1,15 @@
 import * as assert from 'remix/assert'
 import { beforeEach, describe, it } from 'remix/test'
 
-import { verifyPassword } from '../utils/password-hash.ts'
-import { db, resetSocialAuthDatabase, users } from './setup.ts'
+import { db, migrator, seed } from './db.ts'
+import { users } from './data/schema.ts'
+import { verifyPassword } from './utils/password-hash.ts'
 
 beforeEach(async () => {
-  await resetSocialAuthDatabase()
+  await migrator.reset(db, { seed })
 })
 
-describe('social-auth data setup', () => {
+describe('social-auth database seed', () => {
   it('seeds the demo credential users', async () => {
     let admin = await db.findOne(users, { where: { email: 'admin@example.com' } })
     let user = await db.findOne(users, { where: { email: 'user@example.com' } })
