+++ demos/social-auth/app/middleware/database.ts
@@ -1,7 +1,7 @@
 import { Database } from 'remix/data-table'
 import type { Middleware } from 'remix/router'
 
-import { db } from '../data/setup.ts'
+import { db } from '../db.ts'
 
 export function loadDatabase(): Middleware<{
   key: typeof Database
