+++ demos/social-auth/test/helpers.ts
@@ -2,14 +2,13 @@ import { createCookie } from 'remix/cookie'
 import { Cookie, SetCookie } from 'remix/headers'
 import { createMemorySessionStorage } from 'remix/session-storage/memory'
 
+import { db, migrator, seed } from '../app/db.ts'
 import { createSocialAuthRouter, type SocialAuthRouterOptions } from '../app/router.ts'
-import { initializeSocialAuthDatabase, resetSocialAuthDatabase } from '../app/data/setup.ts'
 
 export async function createTestRouter(
   options: Omit<SocialAuthRouterOptions, 'sessionCookie' | 'sessionStorage'> = {},
 ) {
-  await initializeSocialAuthDatabase()
-  await resetSocialAuthDatabase()
+  await migrator.reset(db, { seed })
 
   let sessionCookie = createCookie('social-auth-session', {
     secrets: ['test-social-auth-secret'],
